Split up ObservableMeasurementTableData
This extracts the subclasses from `ObservableMeasurementTableData` to simplify the design and facilitate testing/improving.
Also adds `getHelpText()` to all `MeasurementBuilder` instances, so that it can appear as a tooltip in the selected object summary table.
