+++ qupath-gui-fx/src/main/java/qupath/lib/gui/measure/StringMetadataMeasurementBuilder.java
@@ -1,0 +1,31 @@
+package qupath.lib.gui.measure;
+
+import qupath.lib.objects.PathObject;
+
+class StringMetadataMeasurementBuilder extends AbstractStringMeasurementBuilder {
+
+    private String name;
+
+    StringMetadataMeasurementBuilder(final String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String getName() {
+        return name;
+    }
+
+    @Override
+    public String getHelpText() {
+        return "A metadata value stored within the selected object";
+    }
+
+    @Override
+    public String getMeasurementValue(PathObject pathObject) {
+        if (pathObject != null) {
+            return pathObject.getMetadata().getOrDefault(name, null);
+        }
+        return null;
+    }
+
+}
